submission/requirements.txt
beautifulsoup4==4.6.3
boto==2.49.0
boto3==1.9.40
botocore==1.12.40
bz2file==0.98
certifi==2018.10.15
chardet==3.0.4
cssselect==1.0.3
cycler==0.10.0
docutils==0.14
feedfinder2==0.0.4
feedparser==5.2.1
gensim==3.6.0
idna==2.7
jieba3k==0.35.1
jmespath==0.9.3
kiwisolver==1.0.1
lxml==4.2.5
matplotlib==3.0.1
newspaper3k==0.2.8
nltk==3.3
numpy==1.15.4
pandas==0.23.4
Pillow==5.3.0
pkg-resources==0.0.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.20.1
requests-file==1.4.3
s3transfer==0.1.13
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
sklearn==0.0
smart-open==1.7.1
tinysegmenter==0.3
tldextract==2.2.0
urllib3==1.24.1
